refactor: NewLandingPageをLandingPageにリネーム
