+++ doc/features/public-pages.md
@@ -11,7 +11,7 @@ LPの既存コンテンツを流用し、検索結果に法務ページ以外の
 - `src/routes/_unregistered/demo.flow.tsx` / `src/pages/demo-flow/index.tsx` — 募集から確定通知までのフローデモ
 - `src/routes/_unregistered/demo.shiftboard.tsx` / `src/pages/demo-shift-board/index.tsx` — 店長・シフト担当者向けシフト表デモ
 - `src/components/features/Demo/` — 公開デモ用コンポーネント
-- `src/components/features/NewLandingPage/` — TOPのLP本体、FAQデータ、公開ページ共通フッター
+- `src/components/features/LandingPage/` — TOPのLP本体、FAQデータ、公開ページ共通フッター
 - `src/components/features/LandingPage/FeatureSection.tsx` / `BenefitsSection.tsx` / `FaqSection.tsx` — `/features`・`/faq`で流用している既存LPセクション
 - `src/components/features/ArticleSite/` — Markdown管理の記事サイトとLP記事ミニ導線のソース
 - `scripts/prerender.ts` / `public/sitemap.xml` — 静的HTML生成と検索エンジン向けURL一覧。記事詳細・カテゴリ詳細はMarkdownディレクトリから自動収集する
