+++ src/components/features/LandingPage/HeroSection/index.stories.tsx
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { PricingSection } from ".";
+import { HeroSection } from ".";
 
 const meta = {
-  title: "Features/NewLandingPage/PricingSection",
-  component: PricingSection,
+  title: "Features/LandingPage/HeroSection",
+  component: HeroSection,
   parameters: {
     layout: "fullscreen",
   },
-} satisfies Meta<typeof PricingSection>;
+} satisfies Meta<typeof HeroSection>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
